requirements.txt
fastapi>=0.115.0
fastapi-mcp>=0.1.0
uvicorn>=0.32.0
yfinance>=0.2.49
google-adk>=0.2.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
